Added patchwork layout
According to the documentation, it builds a treemap:

patchwork  draws  the  graph  as  a squarified treemap (see M. Bruls et
al., "Squarified treemaps", Proc. Joint Eurographics and IEEE TCVG  Symp.  on
Visualization,  2000,  pp. 33-42). The clusters of the graph are used to specify the tree.

Also fixed a parsing issue related to a dangling return character in the
output under Linux.